+++ 1.21.4/Common/src/main/java/fuzs/puzzleslib/api/client/event/v1/renderer/ComputeCameraAnglesCallback.java
@@ -13,12 +13,12 @@ public interface ComputeCameraAnglesCallback {
      * Runs before camera angle setup is done, allows for additional control over roll (which vanilla itself does not
      * support) in addition to pitch and yaw.
      *
-     * @param renderer    the game renderer instance
-     * @param camera      the client camera instance
-     * @param partialTick partial ticks
-     * @param yaw         angle for x-axis, transformation is applied second
-     * @param pitch       angle for y-axis, transformation is applied last
-     * @param roll        angle for z-axis, transformation is applied first
+     * @param renderer    the game renderer
+     * @param camera      the client camera
+     * @param partialTick the partial tick
+     * @param yaw         the x-axis angle; transformation is applied second
+     * @param pitch       the y-axis angle; transformation is applied last
+     * @param roll        the z-axis angle; transformation is applied first
      */
     void onComputeCameraAngles(GameRenderer renderer, Camera camera, float partialTick, MutableFloat pitch, MutableFloat yaw, MutableFloat roll);
 }