+++ 1.21.4/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.4.9-1.21.4] - 2025-02-25
+### Added
+- Add `ClientSetupCallback` & `CommonSetupCallback`
+- Add `LootContextKeySetFactory`
+- Add `GiveItemHelper`
+### Changed
+- Update `FabricEventFactory` to handle faulty return values more leniently
+
 ## [v21.4.8-1.21.4] - 2025-02-17
 ### Changed
 - Add `RegistryAccess` to `DataAttachmentRegistry.Builder`